+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 
+## [1.0.3] - 28 June, 2023
+- Add `action` widget list to `BackdropAppBar`
+- Remove the `icon` widget from `SearchBox`
+- Update README.md
+- Update CHANGELOG.md
+- Update pubspec.yaml
+
+
 ## [1.0.2] - 6 June, 2023
 - Add example in plugin